using u32 for token_boundaries

+++ src/compressor/onpair.rs
@@ -12,7 +12,7 @@ pub struct OnPair {
     
     // Dictionary storage  
     dictionary: Vec<u8>,                 // Raw token data
-    token_boundaries: Vec<usize>,        // Token end positions in dictionary
+    token_boundaries: Vec<u32>,        // Token end positions in dictionary
 }
 
 impl OnPair {
@@ -76,7 +76,7 @@ impl OnPair {
             let token = vec![i as u8];
             lpm.insert(&token, i as u16);
             self.dictionary.extend(&token);
-            self.token_boundaries.push(self.dictionary.len());
+            self.token_boundaries.push(self.dictionary.len() as u32);
         }
 
         // Shuffle entries
@@ -113,7 +113,7 @@ impl OnPair {
                     let merged_token = &data[pos - previous_length..pos + match_length];
                     lpm.insert(merged_token, next_token_id);
                     self.dictionary.extend(merged_token);
-                    self.token_boundaries.push(self.dictionary.len());
+                    self.token_boundaries.push(self.dictionary.len() as u32);
                     
                     frequency.remove(&(previous_token_id, match_token_id));
                     previous_token_id = next_token_id;
@@ -197,8 +197,8 @@ impl OnPair {
     /// Buffer must have at least 16 bytes of writable space to avoid undefined behavior.
     #[inline(always)]
     fn decompress_token(&self, token_id: u16, buffer: &mut [u8]) -> usize {
-        let start = self.token_boundaries[token_id as usize];
-        let end = self.token_boundaries[token_id as usize + 1];
+        let start = self.token_boundaries[token_id as usize] as usize;
+        let end = self.token_boundaries[token_id as usize + 1] as usize;
         let token_len = end - start;
 
         unsafe {
@@ -222,7 +222,7 @@ impl OnPair {
     pub fn space_used(&self) -> usize {
         self.compressed_data.len() * std::mem::size_of::<u16>() + 
         self.dictionary.len() + 
-        self.token_boundaries.len() * std::mem::size_of::<usize>()
+        self.token_boundaries.len() * std::mem::size_of::<u32>()
     }
 
     /// Shrinks all internal buffers to fit their current contents
